after optimizing chatroom
