+++ src/navigator/stack/index.js
@@ -3,7 +3,7 @@ import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import Login from '../../modules/Login';
 import OtpScreen from '../../modules/Login/otpScreen';
 import UserProfile from '../../modules/Login/userProfile';
-import {ChatRoom} from '../../modules/Home/Chat/chatRoom';
+import {ChatRoom} from '../../modules/Home/Chat/chatRoom/chatRoom';
 import Home from '../../modules/Home';
 import {string} from '../../utils/strings';
